patch #1745739 - Compiler warning for BSD from Klaus Heinz
git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/trunk@9893 626c5289-ae23-0410-ae9c-e8d60b6d4f22

+++ Source/Swig/misc.c
@@ -293,7 +293,7 @@ String *Swig_string_first_upper(String *s) {
   String *ns = NewStringEmpty();
   char *cs = Char(s);
   if (cs && cs[0] != 0) {
-    Putc(toupper(cs[0]), ns);
+    Putc(toupper((int)cs[0]), ns);
     Append(ns, cs + 1);
   }
   return ns;
@@ -314,7 +314,7 @@ String *Swig_string_first_lower(String *s) {
   String *ns = NewStringEmpty();
   char *cs = Char(s);
   if (cs && cs[0] != 0) {
-    Putc(tolower(cs[0]), ns);
+    Putc(tolower((int)cs[0]), ns);
     Append(ns, cs + 1);
   }
   return ns;
